Fixes and updates

+++ ESPHome/ESPHome-ContactSensor.groovy
@@ -113,7 +113,7 @@ public void parse(Map message) {
 
         case 'state':
             // Check if the entity key matches the message entity key received to update device state
-            if (settings.binarysensor as Integer == message.key) {
+            if (settings.binarysensor as Long == message.key) {
                 String value = message.state ? 'closed' : 'open'
                 if (device.currentValue('motion') != value) {
                     sendEvent([
